+++ app/Models/Event.php
@@ -12,12 +12,12 @@ class Event extends Model
     use Sluggable;
 
     /**
-     * The attributes that are mass assignable.
+     * The attributes that aren't mass assignable.
      *
      * @var array
      */
-    protected $fillable = [
-        'name', 'event_type_id', 'startEvent', 'endEvent', 'startBooking', 'endBooking', 'image_url', 'description', 'dep', 'arr', 'import_only', 'uses_times', 'multiple_bookings_allowed', 'is_oceanic_event',
+    protected $guarded = [
+        'id', 'slug',
     ];
 
     /**