+++ app/Models/EventType.php
@@ -18,13 +18,11 @@ class EventType extends Model
     public $timestamps = false;
 
     /**
-     * The attributes that are mass assignable.
+     * The attributes that aren't mass assignable.
      *
      * @var array
      */
-    protected $fillable = [
-        'id', 'name',
-    ];
+    protected $guarded = ['id'];
 
     protected static $logAttributes = ['*'];
     protected static $logOnlyDirty = true;